Module Colibri2_mappings.Optimizer

val make : unit -> optimize

make () creates a new optimizer.

val push : optimize -> unit

push optimizer pushes a new context level onto the optimizer optimizer.

val pop : optimize -> unit

pop optimizer pops a context level from the optimizer optimizer.

val add : optimize -> Expr.t list -> unit

add optimizer exprs adds the expressions exprs to the optimizer optimizer.

val check : optimize -> [ `Sat | `Unsat | `Unknown ]

check optimizer checks the satisfiability of the optimizer optimizer. Returns `Sat, `Unsat, or `Unknown.

val model : optimize -> model option

model optimizer retrieves the model from the optimizer optimizer, if one exists.

val maximize : optimize -> Expr.t -> handle

maximize optimizer expr maximizes the expression expr in the optimizer optimizer.

val minimize : optimize -> Expr.t -> handle

minimize optimizer expr minimizes the expression expr in the optimizer optimizer.

val interrupt : optimize -> unit

interrupt optimizer interrupts the current optimizer operation.

val get_statistics : optimize -> Statistics.t

get_statistics optimizer retrieves statistics from the optimizer optimizer.