Module S.Bool

val true_ : t

The constant true expression.

val false_ : t

The constant false expression.

val v : bool -> t

v b constructs a Boolean expression from a boolean value.

val not : t -> t

not expr constructs the logical negation of an expression.

val equal : t -> t -> t

equal expr1 expr2 constructs an equality expression.

val distinct : t -> t -> t

distinct expr1 expr2 constructs a distinctness expression.

val and_ : t -> t -> t

and_ expr1 expr2 constructs a logical AND expression.

val or_ : t -> t -> t

or_ expr1 expr2 constructs a logical OR expression.

val implies : t -> t -> t

implies expr1 expr2 constructs a logical implication.

val ite : t -> t -> t -> t

ite cond then_ else_ constructs an if-then-else expression.