Module Smtml.Mappings

include module type of struct include Mappings_intf end

Module Types

module type M = Mappings_intf.M

The M module type defines the core interface for interacting with SMT solvers, including term construction, type handling, and solver interaction.

module type M_with_make = Mappings_intf.M_with_make

The M_with_make module type extends M with a functor for creating instances of M and a flag indicating availability.

module type S = Mappings_intf.S

The S module type defines a simplified interface for interacting with SMT solvers, focusing on model evaluation, solver interaction, and optimization.

module type S_with_fresh = Mappings_intf.S_with_fresh

The S_with_fresh module type extends S with a functor for creating fresh instances of S and a flag indicating availability of the mappings.

module Make (M_with_make : M_with_make) : S_with_fresh