Module Adt.Cons

This module defines the interface for creating ADT constructors.

type t

The abstract type for an ADT constructor definition.

val make : string -> fields:(string * ty option) list -> t

make name ~fields constructs an ADT constructor definition with the given name and a list of fields. Each field in the list is a tuple of its name (string) and its type (ty option). A None type for a field indicates a recursive reference to the ADT being defined.