Module Make.Re

val all : unit -> term

all constructs the regular expression that matches all.

val allchar : unit -> term

allchar constructs the regular expression that matches any character.

val none : unit -> term

empty constructs the empty regular expression.

val star : term -> term

star t constructs the Kleene star of the regular expression term t.

val plus : term -> term

plus t constructs the Kleene plus of the regular expression term t.

val opt : term -> term

opt t constructs the optional regular expression term t.

val comp : term -> term

comp t constructs the complement of the regular expression term t.

val range : term -> term -> term

range t1 t2 constructs a regular expression term matching characters in the range from t1 to t2.

val diff : term -> term -> term

diff t1 t2 constructs the difference of the regular expression terms t1 and t2.

val inter : term -> term -> term

inter t1 t2 constructs the intersection of the regular expression terms t1 and t2.

val loop : min:int -> max:int -> term -> term

loop ~min ~max t constructs a regular expression term matching t repeated between min and max times.

val union : term list -> term

union ts constructs the union of a list of regular expression terms ts.

val concat : term list -> term

concat ts constructs the concatenation of a list of regular expression terms ts.