Module M_with_make.Interp

val to_int : interp -> int

to_int interp converts an interpretation to an integer.

val to_real : interp -> float

to_real interp converts an interpretation to a real number.

val to_bool : interp -> bool

to_bool interp converts an interpretation to a Boolean.

val to_string : interp -> string

to_string interp converts an interpretation to a string.

val to_bitv : interp -> int -> Z.t

to_bitv interp n converts an interpretation to a bitvector of width n.

val to_float : interp -> int -> int -> float

to_float interp e s converts an interpretation to a floating-point number with exponent width e and significand width s.