Module Make.Internals

val name : string

name indicates the name of the solver for query logging and debugging purposes.

val is_available : bool

is_available indicates whether the module is available for use.

val caches_consts : bool

caches_consts indicates whether the solver caches constants.

val was_interrupted : bool Smtml_prelude.ref

was_interrupted indicates if the solver was interrupted.