Module S.Smtlib

val pp : ?name:string -> ?logic:Logic.t -> ?status:[ `Sat | `Unsat | `Unknown ] -> Expr.t list Fmt.t

pp ?name ?logic ?status fmt ts pretty-prints the terms ts in SMT-LIB format using the formatter fmt. Optional parameters include name for the script name, logic for the logic, and status for the script status.