Module Parse.Json

JSON Parsing

val from_string : string -> (t, [> `Msg of string ]) Smtml_prelude.Result.t

from_string s parses a symbol table from a JSON string.

val from_channel : Smtml_prelude.in_channel -> (t, [> `Msg of string ]) Smtml_prelude.Result.t

from_channel ic parses a symbol table from a JSON input channel.

val from_file : Fpath.t -> (t, [> `Msg of string ]) Smtml_prelude.Result.t

from_file path parses a symbol table from a JSON file.