Module Parse.Smtlib

SMT-LIB Parsing

val from_string : string -> (t, [> `Msg of string ]) Smtml_prelude.Result.t

from_string s parses a symbol table from an SMT-LIB string. Warning: Not implemented.

alert unsafe not implemented

val from_channel : Smtml_prelude.in_channel -> (t, [> `Msg of string ]) Smtml_prelude.Result.t

from_channel ic parses a symbol table from an SMT-LIB input channel. Warning: Not implemented.

alert unsafe not implemented

val from_file : Fpath.t -> (t, [> `Msg of string ]) Smtml_prelude.Result.t

from_file path parses a symbol table from an SMT-LIB file. Warning: Not implemented.

alert unsafe not implemented