Module type Optimizer_intf.Intf

The Intf module type defines the interface for creating and working with optimization solvers, including a functor for instantiating solvers and a predefined Z3 solver implementation.

module type S = S

The S module type, which defines the core optimizer interface.

module Make (_ : Mappings_intf.S) : S

Make is a functor that creates an optimizer instance from a given mappings module.

module Z3 : S

Z3 is a predefined optimizer implementation using the Z3 solver.