Module String.Re

type t = regexp expr

val none : t

Regular expression matching nothing.

val all : t

Regular expression matching everything.

val allchar : t

Regular expression matching any single character.

val star : t -> t

star re creates a regular expression matching zero or more occurrences of re.

val plus : t -> t

plus re creates a regular expression matching one or more occurrences of re.

val opt : t -> t

opt re creates a regular expression matching zero or one occurrence of re.

val complement : t -> t

complement re creates a regular expression matching any string NOT matched by re.

val range : string expr -> string expr -> t

range re1 re2 creates a regular expression matching any character in the range from re1 to re2.

val diff : t -> t -> t

diff re1 re2 creates a regular expression matching any string matched by re1 but NOT by re2.

val inter : t -> t -> t

inter re1 re2 creates a regular expression matching any string matched by both re1 and re2.

val loop : min:int -> max:int -> t -> t

loop ~min ~max re creates a regular expression matching re repeated between min and max times.

val union : t list -> t

union res creates a regular expression matching any string matched by any of the regular expressions in res.

val concat : t list -> t

concat res creates a regular expression matching the concatenation of the regular expressions in res.

val pp : t Fmt.t