Module M.Types

val int : ty

int represents the integer type.

val real : ty

real represents the real number type.

val bool : ty

bool represents the Boolean type.

val string : ty

string represents the string type.

val bitv : int -> ty

bitv n represents a bitvector type of width n.

val float : int -> int -> ty

float e s represents a floating-point type with exponent width e and significand width s.

val roundingMode : ty

val regexp : ty

regexp represents the regular expression type.

val ty : term -> ty

ty t retrieves the type of the term t.

val to_ety : ty -> Ty.t

to_ety ty converts the type ty to an smtml type representation.